Finally fixed a compatibility problem with new versions of Visual Studio
Visual Studio 2022 will refuse to compile Sharpamp's GeneralPluginWizard as EnvDTE and other VisualStudio Templates have been significantlly changed from 2008(?), and it results in compilation errors and warnings. I have linked EnvDTE100 in order to compile at this stage, but a re-write for the latest EnvDTE is going to happen some time, and Visual Studio references seem to be more as they were in the past.